+++ linopy/constraints.py
@@ -13,6 +13,7 @@ from typing import Any, Dict, Sequence, Union
 
 import numpy as np
 import pandas as pd
+import polars as pl
 import xarray as xr
 from numpy import array
 from scipy.sparse import csc_matrix
@@ -22,11 +23,15 @@ from linopy import expressions, variables
 from linopy.common import (
     LocIndexer,
     align_lines_by_delimiter,
+    check_has_nulls,
+    check_has_nulls_polars,
+    filter_nulls_polars,
     format_string_as_variable_name,
-    forward_as_properties,
     generate_indices_for_printout,
     get_label_position,
+    group_terms_polars,
     has_optimized_model,
+    infer_schema_polars,
     is_constant,
     maybe_replace_signs,
     print_coord,
@@ -35,6 +40,7 @@ from linopy.common import (
     replace_by_map,
     save_join,
     to_dataframe,
+    to_polars,
 )
 from linopy.config import options
 from linopy.constants import EQUAL, HELPER_DIMS, TERM_DIM, SIGNS_pretty
@@ -62,18 +68,6 @@ def _con_unwrap(con):
     return con.data if isinstance(con, Constraint) else con
 
 
-@forward_as_properties(
-    data=[
-        "attrs",
-        "coords",
-        "indexes",
-        "dims",
-        "sizes",
-    ],
-    labels=["values"],
-    lhs=["nterm"],
-    rhs=["ndim", "shape", "size"],
-)
 class Constraint:
     """
     Projection to a single constraint in a model.
@@ -119,6 +113,89 @@ class Constraint:
 
         self._data = data
         self._model = model
+
+    def __getitem__(self, selector) -> "Constraints":
+        """
+        Get selection from the constraint.
+        This is a wrapper around the xarray __getitem__ method. It returns a
+        new object with the selected data.
+        """
+        data = Dataset({k: self.data[k][selector] for k in self.data}, attrs=self.attrs)
+        return self.__class__(data, self.model, self.name)
+
+    @property
+    def attrs(self):
+        """
+        Get the attributes of the constraint.
+        """
+        return self.data.attrs
+
+    @property
+    def coords(self):
+        """
+        Get the coordinates of the constraint.
+        """
+        return self.data.coords
+
+    @property
+    def indexes(self):
+        """
+        Get the indexes of the constraint.
+        """
+        return self.data.indexes
+
+    @property
+    def dims(self):
+        """
+        Get the dimensions of the constraint.
+        """
+        return self.data.dims
+
+    @property
+    def sizes(self):
+        """
+        Get the sizes of the constraint.
+        """
+        return self.data.sizes
+
+    @property
+    def values(self):
+        """
+        Get the label values of the constraint.
+        """
+        return self.labels.values if self.is_assigned else None
+
+    @property
+    def nterm(self):
+        """
+        Get the number of terms in the constraint.
+        """
+        return self.lhs.nterm
+
+    @property
+    def ndim(self):
+        """
+        Get the number of dimensions of the constraint.
+        """
+        return self.rhs.ndim
+
+    @property
+    def shape(self):
+        """
+        Get the shape of the constraint.
+        """
+        return self.rhs.shape
+
+    @property
+    def size(self):
+        """
+        Get the size of the constraint.
+        """
+        return self.rhs.size
+
+    @property
+    def loc(self):
+        return LocIndexer(self)
 
     @property
     def data(self):
@@ -379,10 +456,6 @@ class Constraint:
         value = DataArray(value).broadcast_like(self.labels)
         self.data["dual"] = value
 
-    @property
-    def shape(self):
-        return self.labels.shape
-
     @classmethod
     def from_rule(cls, model, rule, coords):
         """
@@ -480,14 +553,43 @@ class Constraint:
         agg = dict(coeffs="sum", rhs="first", sign="first")
         agg.update({k: "first" for k in df.columns if k not in agg})
         df = df.groupby(["labels", "vars"], as_index=False).aggregate(agg)
-
-        any_nan = df.isna().any()
-        if any_nan.any():
-            fields = ", ".join("`" + df.columns[any_nan] + "`")
-            raise ValueError(
-                f"Constraint `{self.name}` contains nan's in field(s) {fields}"
-            )
+        check_has_nulls(df, name=f"{self.type} {self.name}")
         return df
+
+    def to_polars(self):
+        """
+        Convert the constraint to a polars DataFrame.
+
+        The resulting DataFrame represents a long table format of the all
+        non-masked constraints with non-zero coefficients. It typically
+        contains the columns `labels`, `coeffs`, `vars`, `rhs`, `sign`.
+
+        Returns
+        -------
+        df : polars.DataFrame
+        """
+        ds = self.data
+
+        keys = [k for k in ds if ("_term" in ds[k].dims) or (k == "labels")]
+        long = to_polars(ds[keys])
+
+        long = filter_nulls_polars(long)
+        long = group_terms_polars(long)
+        check_has_nulls_polars(long, name=f"{self.type} {self.name}")
+
+        short = ds[[k for k in ds if "_term" not in ds[k].dims]]
+        schema = infer_schema_polars(short)
+        schema["sign"] = pl.Enum(["=", "<=", ">="])
+        short = to_polars(short, schema=schema)
+        short = filter_nulls_polars(short)
+        check_has_nulls_polars(short, name=f"{self.type} {self.name}")
+
+        df = pl.concat([short, long], how="diagonal").sort(["labels", "rhs"])
+        # delete subsequent non-null rhs (happens is all vars per label are -1)
+        is_non_null = df["rhs"].is_not_null()
+        prev_non_is_null = is_non_null.shift(1).fill_null(False)
+        df = df.filter(is_non_null & ~prev_non_is_null | ~is_non_null)
+        return df[["labels", "coeffs", "vars", "sign", "rhs"]]
 
     sel = conwrap(Dataset.sel)
 
@@ -591,10 +693,6 @@ class Constraints:
         Remove constraint `name` from the constraints.
         """
         self.data.pop(name)
-
-    @property
-    def loc(self):
-        return LocIndexer(self)
 
     @property
     def labels(self):